+++ b2/data_source_b2_bucket_files_test.go
@@ -69,6 +69,7 @@ func TestAccDataSourceB2BucketFiles_singleFile(t *testing.T) {
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.0.file_id", resourceName, "file_id"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.0.file_info", resourceName, "file_info"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.0.file_name", resourceName, "file_name"),
+					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.0.server_side_encryption", resourceName, "server_side_encryption"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.0.size", resourceName, "size"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.0.upload_timestamp", resourceName, "upload_timestamp"),
 					resource.TestCheckResourceAttr(dataSourceName, "folder_name", ""),
@@ -104,6 +105,7 @@ func TestAccDataSourceB2BucketFiles_multipleFilesWithoutVersions(t *testing.T) {
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.0.file_id", resource2Name, "file_id"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.0.file_info", resource2Name, "file_info"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.0.file_name", resource2Name, "file_name"),
+					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.0.server_side_encryption", resource2Name, "server_side_encryption"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.0.size", resource2Name, "size"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.0.upload_timestamp", resource2Name, "upload_timestamp"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.1.action", resource3Name, "action"),
@@ -114,6 +116,7 @@ func TestAccDataSourceB2BucketFiles_multipleFilesWithoutVersions(t *testing.T) {
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.1.file_id", resource3Name, "file_id"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.1.file_info", resource3Name, "file_info"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.1.file_name", resource3Name, "file_name"),
+					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.0.server_side_encryption", resource3Name, "server_side_encryption"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.1.size", resource3Name, "size"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.1.upload_timestamp", resource3Name, "upload_timestamp"),
 					resource.TestCheckResourceAttr(dataSourceName, "folder_name", ""),
@@ -150,6 +153,7 @@ func TestAccDataSourceB2BucketFiles_multipleFilesWithVersions(t *testing.T) {
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.0.file_id", resource2Name, "file_id"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.0.file_info", resource2Name, "file_info"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.0.file_name", resource2Name, "file_name"),
+					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.0.server_side_encryption", resource2Name, "server_side_encryption"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.0.size", resource2Name, "size"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.0.upload_timestamp", resource2Name, "upload_timestamp"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.1.action", resource1Name, "action"),
@@ -160,6 +164,7 @@ func TestAccDataSourceB2BucketFiles_multipleFilesWithVersions(t *testing.T) {
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.1.file_id", resource1Name, "file_id"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.1.file_info", resource1Name, "file_info"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.1.file_name", resource1Name, "file_name"),
+					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.0.server_side_encryption", resource1Name, "server_side_encryption"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.1.size", resource1Name, "size"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.1.upload_timestamp", resource1Name, "upload_timestamp"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.2.action", resource3Name, "action"),
@@ -171,6 +176,7 @@ func TestAccDataSourceB2BucketFiles_multipleFilesWithVersions(t *testing.T) {
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.2.file_info", resource3Name, "file_info"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.2.file_name", resource3Name, "file_name"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.2.size", resource3Name, "size"),
+					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.0.server_side_encryption", resource3Name, "server_side_encryption"),
 					resource.TestCheckResourceAttrPair(dataSourceName, "file_versions.2.upload_timestamp", resource3Name, "upload_timestamp"),
 					resource.TestCheckResourceAttr(dataSourceName, "folder_name", ""),
 				),
@@ -248,8 +254,12 @@ resource "b2_bucket_file_version" "test3" {
   bucket_id = b2_bucket_file_version.test2.bucket_id
   file_name = "temp2.txt"
   source = b2_bucket_file_version.test2.source
+  server_side_encryption {
+    mode = "SSE-B2"
+    algorithm = "AES256"
+  }
 
-   depends_on = [
+  depends_on = [
     b2_bucket_file_version.test2,
   ]
 }