+++ b2/resource_b2_bucket_file_version_test.go
@@ -42,6 +42,9 @@ func TestAccResourceB2BucketFileVersion_basic(t *testing.T) {
 					resource.TestCheckResourceAttr(resourceName, "content_type", "text/plain"),
 					resource.TestCheckResourceAttr(resourceName, "file_info.%", "0"),
 					resource.TestCheckResourceAttr(resourceName, "file_name", "temp.txt"),
+					resource.TestCheckResourceAttr(resourceName, "server_side_encryption.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "server_side_encryption.0.mode", "none"),
+					resource.TestCheckResourceAttr(resourceName, "server_side_encryption.0.algorithm", ""),
 					resource.TestCheckResourceAttr(resourceName, "size", "5"),
 					resource.TestCheckResourceAttr(resourceName, "source", tempFile),
 					resource.TestMatchResourceAttr(resourceName, "upload_timestamp", regexp.MustCompile("^[0-9]{13}$")),
@@ -74,6 +77,47 @@ func TestAccResourceB2BucketFileVersion_all(t *testing.T) {
 					resource.TestCheckResourceAttr(resourceName, "file_info.%", "1"),
 					resource.TestCheckResourceAttr(resourceName, "file_info.description", "the file"),
 					resource.TestCheckResourceAttr(resourceName, "file_name", "temp.bin"),
+					resource.TestCheckResourceAttr(resourceName, "server_side_encryption.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "server_side_encryption.0.mode", "SSE-B2"),
+					resource.TestCheckResourceAttr(resourceName, "server_side_encryption.0.algorithm", "AES256"),
+					resource.TestCheckResourceAttr(resourceName, "source", tempFile),
+					resource.TestCheckResourceAttr(resourceName, "size", "5"),
+					resource.TestMatchResourceAttr(resourceName, "upload_timestamp", regexp.MustCompile("^[0-9]{13}$")),
+				),
+			},
+		},
+	})
+}
+
+func TestAccResourceB2BucketFileVersion_forceNew(t *testing.T) {
+	parentResourceName := "b2_bucket.test"
+	resourceName := "b2_bucket_file_version.test"
+
+	bucketName := acctest.RandomWithPrefix("test-b2-tfp")
+	tempFile := createTempFile(t, "hello")
+	defer os.Remove(tempFile)
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:          func() { testAccPreCheck(t) },
+		ProviderFactories: providerFactories,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccResourceB2BucketFileVersionConfig_basic(bucketName, tempFile),
+			},
+			{
+				Config: testAccResourceB2BucketFileVersionConfig_all(bucketName, tempFile),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(resourceName, "action", "upload"),
+					resource.TestCheckResourceAttrPair(resourceName, "bucket_id", parentResourceName, "bucket_id"),
+					resource.TestCheckResourceAttr(resourceName, "content_md5", "5d41402abc4b2a76b9719d911017c592"),
+					resource.TestCheckResourceAttr(resourceName, "content_sha1", "aaf4c61ddcc5e8a2dabede0f3b482cd9aea9434d"),
+					resource.TestCheckResourceAttr(resourceName, "content_type", "octet/stream"),
+					resource.TestCheckResourceAttr(resourceName, "file_info.%", "1"),
+					resource.TestCheckResourceAttr(resourceName, "file_info.description", "the file"),
+					resource.TestCheckResourceAttr(resourceName, "file_name", "temp.bin"),
+					resource.TestCheckResourceAttr(resourceName, "server_side_encryption.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "server_side_encryption.0.mode", "SSE-B2"),
+					resource.TestCheckResourceAttr(resourceName, "server_side_encryption.0.algorithm", "AES256"),
 					resource.TestCheckResourceAttr(resourceName, "source", tempFile),
 					resource.TestCheckResourceAttr(resourceName, "size", "5"),
 					resource.TestMatchResourceAttr(resourceName, "upload_timestamp", regexp.MustCompile("^[0-9]{13}$")),
@@ -112,6 +156,10 @@ resource "b2_bucket_file_version" "test" {
   content_type = "octet/stream"
   file_info = {
     description = "the file"
+  }
+  server_side_encryption {
+    mode = "SSE-B2"
+    algorithm = "AES256"
   }
 }
 `, bucketName, tempFile)